requirements.txt
streamlit
pandas
mysql.connector
supabase